chore: refactor with forge settings

+++ src/erc4337-account/ForumAccountFactory.sol
@@ -3,127 +3,109 @@ pragma solidity ^0.8.15;
 
 /* solhint-disable avoid-low-level-calls */
 
-import {Safe, Enum} from '@safe/Safe.sol';
+import {Safe, Enum} from "@safe/Safe.sol";
 
-import {ForumAccount, IEntryPoint} from './ForumAccount.sol';
+import {ForumAccount, IEntryPoint} from "./ForumAccount.sol";
 
 /// @notice Factory to deploy an ERC4337 Account
 /// @author ForumDAOs (https://forumdaos.com)
 contract ForumAccountFactory {
-	/// ----------------------------------------------------------------------------------------
-	/// Errors and Events
-	/// ----------------------------------------------------------------------------------------
+    /// ----------------------------------------------------------------------------------------
+    /// Errors and Events
+    /// ----------------------------------------------------------------------------------------
 
-	error NullDeploy();
+    error NullDeploy();
 
-	error AccountInitFailed();
+    error AccountInitFailed();
 
-	/// ----------------------------------------------------------------------------------------
-	/// Factory Storage
-	/// ----------------------------------------------------------------------------------------
+    /// ----------------------------------------------------------------------------------------
+    /// Factory Storage
+    /// ----------------------------------------------------------------------------------------
 
-	// Template contract to use for new individual ERC4337 accounts
-	ForumAccount public immutable forumAccountSingleton;
+    // Template contract to use for new individual ERC4337 accounts
+    ForumAccount public immutable forumAccountSingleton;
 
-	// Entry point used for ERC4337 accounts
-	address public immutable entryPoint;
+    // Entry point used for ERC4337 accounts
+    address public immutable entryPoint;
 
-	// Fallback handler for Gnosis Safe
-	address public immutable gnosisFallbackLibrary;
+    // Fallback handler for Gnosis Safe
+    address public immutable gnosisFallbackLibrary;
 
-	// Address of deterministic deployment proxy, allowing address generation independent of deployer or nonce
-	// https://github.com/Arachnid/deterministic-deployment-proxy
-	address public constant DETERMINISTIC_DEPLOYMENT_PROXY =
-		0x4e59b44847b379578588920cA78FbF26c0B4956C;
+    // Address of deterministic deployment proxy, allowing address generation independent of deployer or nonce
+    // https://github.com/Arachnid/deterministic-deployment-proxy
+    address public constant DETERMINISTIC_DEPLOYMENT_PROXY = 0x4e59b44847b379578588920cA78FbF26c0B4956C;
 
-	bytes private _createProxyData;
+    bytes private _createProxyData;
 
-	bytes32 private immutable _createProxyDataHash;
+    bytes32 private immutable _createProxyDataHash;
 
-	/// ----------------------------------------------------------------------------------------
-	/// Constructor
-	/// ----------------------------------------------------------------------------------------
+    /// ----------------------------------------------------------------------------------------
+    /// Constructor
+    /// ----------------------------------------------------------------------------------------
 
-	constructor(
-		ForumAccount _forumAccountSingleton,
-		address _entryPoint,
-		address _gnosisFallbackLibrary
-	) {
-		forumAccountSingleton = _forumAccountSingleton;
+    constructor(ForumAccount _forumAccountSingleton, address _entryPoint, address _gnosisFallbackLibrary) {
+        forumAccountSingleton = _forumAccountSingleton;
 
-		entryPoint = _entryPoint;
+        entryPoint = _entryPoint;
 
-		gnosisFallbackLibrary = _gnosisFallbackLibrary;
+        gnosisFallbackLibrary = _gnosisFallbackLibrary;
 
-		// Data sent to the deterministic deployment proxy to deploy a new ERC4337 account
-		_createProxyData = abi.encodePacked(
-			// constructor
-			bytes10(0x3d602d80600a3d3981f3),
-			// proxy code
-			bytes10(0x363d3d373d3d3d363d73),
-			forumAccountSingleton,
-			bytes15(0x5af43d82803e903d91602b57fd5bf3)
-		);
+        // Data sent to the deterministic deployment proxy to deploy a new ERC4337 account
+        _createProxyData = abi.encodePacked(
+            // constructor
+            bytes10(0x3d602d80600a3d3981f3),
+            // proxy code
+            bytes10(0x363d3d373d3d3d363d73),
+            forumAccountSingleton,
+            bytes15(0x5af43d82803e903d91602b57fd5bf3)
+        );
 
-		// Hashed so we can save this as immutable to save gas calcualting address
-		_createProxyDataHash = keccak256(_createProxyData);
-	}
+        // Hashed so we can save this as immutable to save gas calcualting address
+        _createProxyDataHash = keccak256(_createProxyData);
+    }
 
-	/// ----------------------------------------------------------------------------------------
-	/// Factory Logic
-	/// ----------------------------------------------------------------------------------------
+    /// ----------------------------------------------------------------------------------------
+    /// Factory Logic
+    /// ----------------------------------------------------------------------------------------
 
-	/**
-	 * @notice Deploys a new ERC4337 account built on a Gnosis safe
-	 * @param owner Public key for secp256r1 signer
-	 * @return account The deployed account
-	 * @dev Returns an existing account address so that entryPoint.getSenderAddress() works even after account creation
-	 */
-	function createForumAccount(
-		uint256[2] calldata owner
-	) external payable virtual returns (address payable account) {
-		bytes32 accountSalt = keccak256(abi.encodePacked(owner));
+    /**
+     * @notice Deploys a new ERC4337 account built on a Gnosis safe
+     * @param owner Public key for secp256r1 signer
+     * @return account The deployed account
+     * @dev Returns an existing account address so that entryPoint.getSenderAddress() works even after account creation
+     */
+    function createForumAccount(uint256[2] calldata owner) external payable virtual returns (address payable account) {
+        bytes32 accountSalt = keccak256(abi.encodePacked(owner));
 
-		address addr = getAddress(accountSalt);
-		uint codeSize = addr.code.length;
-		if (codeSize > 0) {
-			return payable(addr);
-		}
+        address addr = getAddress(accountSalt);
+        uint256 codeSize = addr.code.length;
+        if (codeSize > 0) {
+            return payable(addr);
+        }
 
-		// Deploy the account determinstically based on the salt
-		(bool successCreate, bytes memory responseCreate) = DETERMINISTIC_DEPLOYMENT_PROXY.call{
-			value: 0
-		}(abi.encodePacked(accountSalt, _createProxyData));
+        // Deploy the account determinstically based on the salt
+        (bool successCreate, bytes memory responseCreate) =
+            DETERMINISTIC_DEPLOYMENT_PROXY.call{value: 0}(abi.encodePacked(accountSalt, _createProxyData));
 
-		// If successful, convert response to address to be returned
-		account = payable(address(uint160(bytes20(responseCreate))));
+        // If successful, convert response to address to be returned
+        account = payable(address(uint160(bytes20(responseCreate))));
 
-		if (!successCreate || account == address(0)) revert NullDeploy();
+        if (!successCreate || account == address(0)) revert NullDeploy();
 
-		ForumAccount(payable(account)).initialize(entryPoint, owner, gnosisFallbackLibrary);
-	}
+        ForumAccount(payable(account)).initialize(entryPoint, owner, gnosisFallbackLibrary);
+    }
 
-	/// ----------------------------------------------------------------------------------------
-	/// Factory Internal
-	/// ----------------------------------------------------------------------------------------
+    /// ----------------------------------------------------------------------------------------
+    /// Factory Internal
+    /// ----------------------------------------------------------------------------------------
 
-	/**
-	 * @notice Get the address of an account that would be returned by createAccount()
-	 * @dev Salt should be keccak256(abi.encodePacked(owner)) where owner is the [x, y] public key for secp256r1 signer
-	 */
-	function getAddress(bytes32 salt) public view returns (address clone) {
-		return
-			address(
-				bytes20(
-					keccak256(
-						abi.encodePacked(
-							bytes1(0xff),
-							DETERMINISTIC_DEPLOYMENT_PROXY,
-							salt,
-							_createProxyDataHash
-						)
-					) << 96
-				)
-			);
-	}
+    /**
+     * @notice Get the address of an account that would be returned by createAccount()
+     * @dev Salt should be keccak256(abi.encodePacked(owner)) where owner is the [x, y] public key for secp256r1 signer
+     */
+    function getAddress(bytes32 salt) public view returns (address clone) {
+        return address(
+            bytes20(keccak256(abi.encodePacked(bytes1(0xff), DETERMINISTIC_DEPLOYMENT_PROXY, salt, _createProxyDataHash)) << 96)
+        );
+    }
 }
