+++ src/libraries/HexToLiteralBytes.sol
@@ -5,28 +5,28 @@ pragma solidity ^0.8.15;
  * @notice Library to convert hexadecimal strings to raw bytes
  */
 library HexToLiteralBytes {
-	// Convert an hexadecimal string to raw bytes
-	function fromHex(string memory s) internal pure returns (bytes memory) {
-		bytes memory ss = bytes(s);
-		require(ss.length % 2 == 0, 'hex length not even');
-		bytes memory r = new bytes(ss.length / 2);
-		for (uint i = 0; i < ss.length / 2; ++i) {
-			r[i] = bytes1(fromHexChar(uint8(ss[2 * i])) * 16 + fromHexChar(uint8(ss[2 * i + 1])));
-		}
-		return r;
-	}
+    // Convert an hexadecimal string to raw bytes
+    function fromHex(string memory s) internal pure returns (bytes memory) {
+        bytes memory ss = bytes(s);
+        require(ss.length % 2 == 0, "hex length not even");
+        bytes memory r = new bytes(ss.length / 2);
+        for (uint256 i = 0; i < ss.length / 2; ++i) {
+            r[i] = bytes1(fromHexChar(uint8(ss[2 * i])) * 16 + fromHexChar(uint8(ss[2 * i + 1])));
+        }
+        return r;
+    }
 
-	// Convert an hexadecimal character to their value
-	function fromHexChar(uint8 c) internal pure returns (uint8) {
-		if (bytes1(c) >= bytes1('0') && bytes1(c) <= bytes1('9')) {
-			return c - uint8(bytes1('0'));
-		}
-		if (bytes1(c) >= bytes1('a') && bytes1(c) <= bytes1('f')) {
-			return 10 + c - uint8(bytes1('a'));
-		}
-		if (bytes1(c) >= bytes1('A') && bytes1(c) <= bytes1('F')) {
-			return 10 + c - uint8(bytes1('A'));
-		}
-		revert('failed hex conversion');
-	}
+    // Convert an hexadecimal character to their value
+    function fromHexChar(uint8 c) internal pure returns (uint8) {
+        if (bytes1(c) >= bytes1("0") && bytes1(c) <= bytes1("9")) {
+            return c - uint8(bytes1("0"));
+        }
+        if (bytes1(c) >= bytes1("a") && bytes1(c) <= bytes1("f")) {
+            return 10 + c - uint8(bytes1("a"));
+        }
+        if (bytes1(c) >= bytes1("A") && bytes1(c) <= bytes1("F")) {
+            return 10 + c - uint8(bytes1("A"));
+        }
+        revert("failed hex conversion");
+    }
 }
